% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hydroMet.R
\docType{class}
\name{hydroMet-class}
\alias{hydroMet-class}
\alias{hydroMet}
\title{\code{hydroMet} superclass object}
\value{
A basic hydroMet class object.
}
\description{
A suitable object for store basic information about an hydro-meteorological station.
}
\section{Slots}{

\describe{
\item{\code{id}}{numeric. This is the ID assigned by the agency.}

\item{\code{agency}}{character. The name of the agency (or institution) that provides the data of the station.}

\item{\code{station}}{character. The name of the (hydro)-meteorological station.}

\item{\code{lat}}{numeric. Latitude of the station.}

\item{\code{long}}{numeric. Longitude of the station}

\item{\code{alt}}{numeric. Altitude of the station.}

\item{\code{country}}{character. Country where the station is located. Argentina is set as default value.}

\item{\code{province}}{character. Name of the province where the station is located. Mendoza is set as default value.}

\item{\code{river}}{character. Basin river's name.}

\item{\code{active}}{logical. It indicates whether or not the station is currently operated. Default value is \code{TRUE}.}
}}

