% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hydroMet_BDHI.R
\docType{class}
\name{hydroMet_BDHI-class}
\alias{hydroMet_BDHI-class}
\alias{hydroMet_BDHI}
\title{\code{hydroMet} subclass for BDHI (Base de Datos Hidrologica Integrada) data}
\value{
A hydroMet_BDHI class object.
}
\description{
An suitable object for store hydro-meteorological data from BDHI.
}
\section{Slots}{

\describe{
\item{\code{Qmd}}{data.frame from \link{read_BDHI} containing daily mean river discharge [m3/s].}

\item{\code{Qmm}}{data.frame from \link{read_BDHI} containing monthly mean river discharge [m3/s].}

\item{\code{precip}}{data.frame from \link{read_BDHI} containing daily liquid precipitation [mm].}

\item{\code{tdb}}{data.frame from \link{read_BDHI} containing subdaily dry bulb temperature [ºC].}

\item{\code{tmax}}{data.frame from \link{read_BDHI} containing daily maximum air temperature [ºC].}

\item{\code{tmin}}{data.frame from \link{read_BDHI} containing daily minimum air temperature [ºC].}

\item{\code{swe}}{data.frame from \link{read_BDHI} containing daily snow water equivalent [mm].}

\item{\code{hr}}{data.frame from \link{read_BDHI} containing subdaily relative humidity [\%].}

\item{\code{wspd}}{data.frame from \link{read_BDHI} containing subdaily wind speed [km/hr].}

\item{\code{wdir}}{data.frame from \link{read_BDHI} containing subdaily wind direction [º].}

\item{\code{evap}}{data.frame from \link{read_BDHI} containing daily pan-evaporation [mm].}

\item{\code{anem}}{data.frame from \link{read_BDHI} containing daily wind speed above the evap tank [km/hr].}

\item{\code{patm}}{data.frame from \link{read_BDHI} containing subdaily atmospheric pressure [mbar].}
}}

