% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hydroMet_compact.R
\docType{class}
\name{hydroMet_compact-class}
\alias{hydroMet_compact-class}
\alias{hydroMet_compact}
\title{\code{hydroMet} subclass for compact data}
\value{
A hydroMet_compact class object.
}
\description{
This subclass is useful for storing in a single data frame ready to use hydro-meteorological series or many variables of the same kind (e.g. lets say precipitacion series).
}
\section{Slots}{

\describe{
\item{\code{compact}}{data.frame with Date as first column (class 'Date' or 'POSIXct'). All other columns are the numeric hydro-meteorological variables (double). This subclass was though to join in a single table ready to use data (e.g. in modelling). You can also use it to put together variables of the same kind (e.g. precipitation records) to make some regional analysis.}
}}

