% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_hydroMet_IANIGLA.R
\docType{class}
\name{hydroMet_IANIGLA-class}
\alias{hydroMet_IANIGLA-class}
\alias{hydroMet_IANIGLA}
\title{\code{hydroMet} subclass for IANIGLA (Instituto Argentino de Nivología, Glaciología y Ciencias Ambientales) data}
\value{
A hydroMet_IANIGLA class object.
}
\description{
A suitable object for store hydro-meteorological data provided by IANIGLA.
}
\section{Slots}{

\describe{
\item{\code{date}}{time serie of dates (class \code{POSIXct} or \code{Date}).}

\item{\code{tair}}{numeric matrix with air temperature.}

\item{\code{hr}}{numeric matrix with relative humidity.}

\item{\code{patm}}{numeric matrix with atmospheric pressure.}

\item{\code{precip}}{numeric matrix with precipitacion.}

\item{\code{wspd}}{numeric matrix with wind speed.}

\item{\code{wdir}}{numeric matrix with wind direction.}

\item{\code{kin}}{numeric matrix with incoming short-wave radiation.}

\item{\code{hsnow}}{numeric matrix with snow height.}

\item{\code{tsoil}}{numeric matrix with soil temperature.}

\item{\code{hwat}}{numeric matrix with stream water level.}
}}

